+++ apps/example/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @bazokhan/sightglass-example
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @bazokhan/sightglass-core@0.1.2
+  - @bazokhan/sightglass-express@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes